Added Skybox rendering

+++ Hazard-Utility/src/Utility/StringUtil.h
@@ -13,7 +13,7 @@ public:
 	static std::string RemoveAtEnd(std::string& source, char val);
 	static size_t OffsetOf(const std::string& source, char delim, size_t startOffset = 0);
 	static size_t OffsetOf(const std::string& source, const std::string& value, size_t startOffset = 0);
-	static std::string BytesToString(const size_t& bytes);
+	static std::string BytesToString(const int64_t& bytes);
 	static std::string Replace(const std::string& value, const std::string& find, const std::string& replaceWith);
 	static std::string& ToLower(std::string& string);
 	static bool IsMatching(const std::string& value, const std::string& compareTo, bool caseSensitive = false, bool stripWhiteSpaces = true, bool stripUnderScores = true);
